backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
pandas==2.2.0
pytz==2024.1
